Move Bands to bands sub-module
