+++ fooof/bands/__init__.py
@@ -1,0 +1,3 @@
+"""Bands sub-module for FOOOF."""
+
+from .bands import Bands
