+++ fooof/data/__init__.py
@@ -1,4 +1,3 @@
 """Data sub-module for FOOOF."""
 
-from .bands import Bands
 from .data import FOOOFSettings, FOOOFMetaData, FOOOFResults, SimParams
