+++ fooof/tests/bands/test_bands.py
@@ -2,7 +2,7 @@
 
 from py.test import raises
 
-from fooof.data.bands import *
+from fooof.bands.bands import *
 
 ###################################################################################################
 ###################################################################################################
